requirements.txt
aifs==0.0.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
editor==1.6.6
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.3
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
inquirer==3.2.4
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
litellm==1.37.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
MouseInfo==0.1.3
multidict==6.0.5
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
open-interpreter==0.2.5
openai==1.28.1
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
posthog==3.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
PyYAML==6.0.1
pyzmq==26.0.3
readchar==4.0.6
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rubicon-objc==0.4.9
runs==1.2.2
screeninfo==0.8.1
selenium==4.20.0
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
tiktoken==0.6.0
tokenizers==0.19.1
tokentrim==0.1.13
toml==0.10.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wget==3.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.9.4
zipp==3.18.1
